<Setup Lab Demo and Created Pill Organiser Proof of Concept> Retooled Troubler System to now use Lab based machinery in preparation for another demo, but this one uses the full Troubler System and is not scripted.
Also added a feature to TroubleTask where an image can be specified to be used for that Task if it is being presented to user as a possible task(is this your task? y/n) and another image to be displayed when that task is used to create an alert.

Also created a new activity PillsActivity that serves as proof of concept for a medication organising app, using nfc to trigger it.