<Implemented Early Version of the Alert tracking into Troubler> Implemented an intial version of a system to track and read out alerts to employee's based on their id. Currently alerts are created through the raise trouble ticket dialog tree and are wiped as soon as the activity closes, while Employee records as well as what type of employee is assigned to what task is hard coded. i.e. all alerts are currently assigned to employee's of type Plumber, of which only one exists, and who name and id are hard coded into the TroubleTicketActivity.
In addition, I've implemented some features dan requested for a demo, specifically, unique instructions for specific alerts and a dialog exchange when the user has a fume hood alert read out to them, instead of reading out the alert, a dialog starts asking about a gas sensor which then leads back into the usual TroublerStart Dialog. However this system is hard coded into the normal dialog and alert reading system and has been write in a horrible way and should only be used for the demo and not for a future implementation of this system for overriding alert description with unique responses under specific circumstances.